multiload: use guint64 as data source for graphs

+++ multiload/src/linux-proc.h
@@ -3,11 +3,11 @@
 
 #include <load-graph.h>
 
-G_GNUC_INTERNAL void GetLoad     (int Maximum, int data [cpuload_n],  LoadGraph *g);
-G_GNUC_INTERNAL void GetDiskLoad (int Maximum, int data [diskload_n], LoadGraph *g);
-G_GNUC_INTERNAL void GetMemory   (int Maximum, int data [memload_n],  LoadGraph *g);
-G_GNUC_INTERNAL void GetSwap     (int Maximum, int data [swapload_n], LoadGraph *g);
-G_GNUC_INTERNAL void GetLoadAvg  (int Maximum, int data [2],          LoadGraph *g);
-G_GNUC_INTERNAL void GetNet      (int Maximum, int data [4],          LoadGraph *g);
+G_GNUC_INTERNAL void GetLoad     (guint64 Maximum, guint64 data [cpuload_n],  LoadGraph *g);
+G_GNUC_INTERNAL void GetDiskLoad (guint64 Maximum, guint64 data [diskload_n], LoadGraph *g);
+G_GNUC_INTERNAL void GetMemory   (guint64 Maximum, guint64 data [memload_n],  LoadGraph *g);
+G_GNUC_INTERNAL void GetSwap     (guint64 Maximum, guint64 data [swapload_n], LoadGraph *g);
+G_GNUC_INTERNAL void GetLoadAvg  (guint64 Maximum, guint64 data [2],          LoadGraph *g);
+G_GNUC_INTERNAL void GetNet      (guint64 Maximum, guint64 data [4],          LoadGraph *g);
 
 #endif
